pyproject.toml
[project]
name = "repl-nix-videosummarizergemini"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "google-api-python-client>=2.151.0",
    "youtube-transcript-api>=0.6.2",
    "google-generativeai>=0.8.3",
]
